+++ maslevtsov.stanislav/T3/count.hpp
@@ -1,0 +1,10 @@
+#ifndef COUNT_HPP
+#define COUNT_HPP
+
+#include "shapes.hpp"
+
+namespace maslevtsov {
+  void count_vertexes(const std::vector< Polygon >& polygons, std::istream& in, std::ostream& out);
+}
+
+#endif
