+++ maslevtsov.stanislav/T3/get_area.hpp
@@ -1,0 +1,10 @@
+#ifndef GET_AREA_HPP
+#define GET_AREA_HPP
+
+#include "shapes.hpp"
+
+namespace maslevtsov {
+  void get_area(const std::vector< Polygon >& polygons, std::istream& in, std::ostream& out);
+}
+
+#endif
